+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All Notable changes to `RecursivePagination` will be documented in this file
 
+## Unreleased 0.1.0-alpha.2
+
+### Fixed
+- `RecursivePagination::getNextPage()` bug fixed: #19 issue
+
+### Removed
+- First parameter of `RecursivePagination::getNextPage()` method has been removed
+
 ## 0.1.0-alpha.1 [2016-05-27]
 
 ### Added